refactor: Merge `lock-manager.ts` into `cache/index.ts`
